Rethink some of the structure
